requirements.txt
Flask
gunicorn
python-dotenv
google-generativeai
google-api-python-client
